requirements/base.txt
Django==2.2
psycopg2-binary
djangorestframework
django-filter
django-rest-swagger
django-cors-headers